+++ packages/db-postgres/src/deleteMany.ts
@@ -12,7 +12,7 @@ export const deleteMany: DeleteMany = async function deleteMany(
   this: PostgresAdapter,
   { collection, req = {} as PayloadRequest, where },
 ) {
-  const db = this.sessions[req.transactionID]?.db || this.drizzle
+  const db = this.sessions[await req.transactionID]?.db || this.drizzle
   const collectionConfig = this.payload.collections[collection].config
 
   const tableName = this.tableNameMap.get(toSnakeCase(collectionConfig.slug))