+++ packages/db-postgres/src/deleteOne.ts
@@ -15,7 +15,7 @@ export const deleteOne: DeleteOne = async function deleteOne(
   this: PostgresAdapter,
   { collection: collectionSlug, req = {} as PayloadRequest, where: whereArg },
 ) {
-  const db = this.sessions[req.transactionID]?.db || this.drizzle
+  const db = this.sessions[await req.transactionID]?.db || this.drizzle
   const collection = this.payload.collections[collectionSlug].config
 
   const tableName = this.tableNameMap.get(toSnakeCase(collection.slug))