+++ test/field-error-states/collections/ValidateDraftsOff/index.ts
@@ -1,0 +1,12 @@
+import type { CollectionConfig } from '../../../../packages/payload/src/collections/config/types'
+
+import { slugs } from '../../shared'
+import { ValidateDraftsOn } from '../ValidateDraftsOn'
+
+export const ValidateDraftsOff: CollectionConfig = {
+  ...ValidateDraftsOn,
+  slug: slugs.validateDraftsOff,
+  versions: {
+    drafts: true,
+  },
+}